requirements.txt
gradio
mysql-connector-python